refactor: move mode to its on filw
