+++ apps/server/app/routes/_dashboard.tsx
@@ -1,15 +1,14 @@
-import {redirect, type LoaderArgs, type V2_MetaFunction} from '@remix-run/node'
-import {Outlet, useLoaderData, useLocation, useNavigate} from '@remix-run/react'
+import {type LoaderArgs, type V2_MetaFunction} from '@remix-run/node'
+import {Outlet, useLoaderData} from '@remix-run/react'
 
+import {EnvSelector} from '~/components/env-selector'
 import {Icons} from '~/components/icons'
 import {MainNav} from '~/components/main-nav'
 import {Search} from '~/components/search'
 import TeamSwitcher from '~/components/team-switcher'
-import {Tabs, TabsList, TabsTrigger} from '~/components/ui/tabs'
 import {UserNav} from '~/components/user-nav'
 import {requireUserId} from '~/utils/auth.server'
-import {cn} from '~/utils/css'
-import {env} from '~/utils/env.server'
+import {getEnabledModes, requireEnabledMode} from '~/utils/mode.server'
 
 export const meta: V2_MetaFunction = () => {
   return [{title: 'Dashboard'}]
@@ -51,131 +50,5 @@ export default function DashboardLayout() {
         <Outlet />
       </div>
     </>
-  )
-}
-
-type Mode = keyof ReturnType<typeof getEnabledModes>
-
-/**
- * Get enabled mode from environment variables
- */
-function getEnabledModes() {
-  return {
-    dev: env.DEV_MODE_ENABLED,
-    test: !!env.TESTNET_BITCOIN_CORE_URL,
-    production: !!env.MAINNET_BITCOIN_CORE_URL,
-  }
-}
-
-/**
- * Check if mode is enabled and return it
- *
- * @param userMode the mode user wants to use
- * @returns
- */
-function requireEnabledMode(userMode: string = '') {
-  const enabledModes = getEnabledModes()
-  let mode: Mode
-
-  switch (userMode) {
-    case 'dev':
-      mode = 'dev'
-      break
-    case 'test':
-      mode = 'test'
-      break
-    case '':
-      mode = 'production'
-      break
-    default:
-      throw new Response(`Invalid mode: ${userMode}`, {
-        status: 400,
-        statusText: 'Invalid mode',
-      })
-  }
-
-  // if mode is not enabled, redirect to the first enabled mode
-  if (!enabledModes[mode]) {
-    const priority: Mode[] = ['dev', 'production', 'test']
-    // find the first enabled mode
-    mode = priority.filter(m => enabledModes[m])[0]
-
-    throw redirect(mode === 'production' ? '/' : `/${mode}`, {
-      statusText: `Mode ${mode} is not enabled`,
-    })
-  }
-
-  return mode
-}
-
-function EnvSelector({
-  current,
-  modes,
-}: {
-  current: Mode
-  modes: ReturnType<typeof getEnabledModes>
-}) {
-  const navigate = useNavigate()
-  // get current route
-  const {pathname} = useLocation()
-  // remove mode from route
-  const route = pathname.replace(/^\/(dev|test)\//, '/')
-
-  return (
-    <Tabs
-      value={current}
-      onValueChange={env =>
-        navigate(`${env === 'production' ? '' : env}${route}`)
-      }
-      className={cn('flex border-b -mt-[22px] -mb-px', {
-        'border-b-info': current === 'dev',
-        'border-b-warning': current === 'test',
-      })}
-    >
-      <TabsList
-        className={cn('mx-auto rounded-b-none rounded-t-sm h-auto', {
-          'bg-info': current === 'dev',
-          'bg-warning': current === 'test',
-        })}
-      >
-        {modes.dev && (
-          <EnvTabTrigger current={current} value="dev">
-            Development
-          </EnvTabTrigger>
-        )}
-        {modes.test && (
-          <EnvTabTrigger current={current} value="test">
-            Test
-          </EnvTabTrigger>
-        )}
-        {modes.production && (
-          <EnvTabTrigger current={current} value="production">
-            Production
-          </EnvTabTrigger>
-        )}
-      </TabsList>
-    </Tabs>
-  )
-}
-
-function EnvTabTrigger({
-  current,
-  value,
-  children,
-}: {
-  current: Mode
-  value: Mode
-  children: React.ReactNode
-}) {
-  return (
-    <TabsTrigger
-      className={cn(`text-[10px] leading-[14px] px-1.5 py-0 rounded-[2px]`, {
-        'text-info-foreground': current === 'dev',
-        'text-warning-foreground': current === 'test',
-      })}
-      value={value}
-    >
-      {children}
-    </TabsTrigger>
   )
 }
